Separate tests into to a single folder
