+++ .github/workflows/ci.yml
@@ -219,8 +219,8 @@ jobs:
       - name: Visual Regression Tests
         id: visualregressiontests
         run: |
-          cd samples/StaticHtmlComposites
-          rushx test:e2e
+          cd samples/tests
+          rushx test:e2e:bundle
         env:
           CONNECTION_STRING: ${{ secrets.CONNECTION_STRING }}
       - name: Upload snapshot diff
@@ -228,7 +228,7 @@ jobs:
         uses: actions/upload-artifact@v2
         with:
           name: snapshots
-          path: samples/StaticHtmlComposites/test-results/ # or path/to/artifact
+          path: samples/tests/test-results/ # or path/to/artifact
       - name: Comment on PR
         if: ${{ github.event_name == 'pull_request' && always() && steps.visualregressiontests.outcome == 'failure' && !contains( github.event.pull_request.labels.*.name, 'ui change') }}
         uses: actions/github-script@v3
