+++ .github/workflows/update-snapshot.yml
@@ -57,8 +57,8 @@ jobs:
         run: rush build -t sample-static-html-composites
       - name: Update snapshot
         run: |
-          cd samples/StaticHtmlComposites
-          rushx update:snapshot
+          cd samples/tests
+          rushx update:snapshot:bundle
         env:
           CONNECTION_STRING: ${{ secrets.CONNECTION_STRING }}
       # Check if files have changed
@@ -76,7 +76,7 @@ jobs:
       - name: Commit new snapshots
         if: ${{ steps.changescheck.outputs.hasChanged == 'true' }}
         run: |
-          git add samples/StaticHtmlComposites
+          git add samples/tests/StaticHtmlComposites
           git commit -m 'Update js bundle automation snapshots'
           git push
