+++ samples/Calling/package.json
@@ -58,7 +58,7 @@
     "babel-loader": "8.1.0",
     "concurrently": "^5.3.0",
     "copyfiles": "^2.4.1",
-    "cross-env": "^5.2.0",
+    "cross-env": "~7.0.3",
     "css-loader": "~4.3.0",
     "eslint": "^7.7.0",
     "eslint-config-prettier": "^6.12.0",
