+++ samples/tests/.env
@@ -1,0 +1,1 @@
+CONNECTION_STRING=[REPLACE_WITH_CONNECTION_STRING]
